refactored separate list components with a single TrackerList component
